Move libarchive stuff from NDSCart to Platform
