+++ src/NDSCart.cpp
@@ -18,8 +18,6 @@
 
 #include <stdio.h>
 #include <string.h>
-#include <archive.h>
-#include <archive_entry.h>
 #include "NDS.h"
 #include "NDSCart.h"
 #include "ARM.h"
@@ -867,82 +865,28 @@ bool ReadROMParams(u32 gamecode, u32* params)
     }
 }
 
-bool CheckArchiveExtensions(const char *ext)
-{
-    const char *extensions[] =
-    {
-        "zip", ".7z", "lz4"
-    };
-
-    for(int i = 0; i < sizeof(extensions)/sizeof(char *); ++i)
-        if(!strcasecmp(ext, extensions[i]))
-            return true;
-
-    return false;
-}
-
 bool LoadROM(const char* path, const char* sram, bool direct)
 {
     // TODO: streaming mode? for really big ROMs or systems with limited RAM
     // for now we're lazy
 
     const char* ext = &path[strlen(path)-3];
-    bool IsArchive = CheckArchiveExtensions(ext);
+    bool IsArchive = Platform::CheckArchiveExtensions(ext);
+
     FILE* f;
-    if(!IsArchive)
+    u8 *extractBuf; size_t extractSize;
+
+    if(IsArchive)
+    {
+        extractBuf = Platform::OpenArchive(path, &extractSize);
+        if(!extractBuf)
+            return false;
+    }
+    else
     {
         f = Platform::OpenFile(path, "rb");
         if (!f)
         {
-            return false;
-        }
-    }
-
-    u8 *extractBuf = nullptr; size_t extractSize;
-    if(IsArchive)
-    {
-        struct archive *a = archive_read_new();
-        struct archive_entry *entry;
-
-        archive_read_support_filter_all(a);
-        archive_read_support_format_all(a);
-        int r = archive_read_open_filename(a, path, 10240);
-        if (r != ARCHIVE_OK)
-        {
-            printf("libarchive error : '%s'\n", archive_error_string(a));
-            return false;
-        }
-
-        while(archive_read_next_header(a, &entry) == ARCHIVE_OK)
-        {
-            ext = archive_entry_pathname(entry);
-            printf("Archive Entry = %s\n", ext);
-            ext = &ext[strlen(ext)-3];
-            if(!strcasecmp(ext, "nds") || !strcasecmp(ext, "srl"))
-            {
-                extractSize = archive_entry_size(entry);
-                printf("NDS Rom Size = %lu\n", extractSize);
-                extractBuf = new u8[extractSize];
-                ssize_t bytes_read = archive_read_data(a, extractBuf, extractSize);
-                if(bytes_read < 0)
-                {
-                    printf("Error in archive_read_data!\n");
-                    return false;
-                }
-                break;
-            }
-        }
-
-        if(!extractBuf)
-        {
-            printf("No NDS rom found in archive!\n");
-            return false;
-        }
-
-        r = archive_read_free(a);
-        if(r != ARCHIVE_OK)
-        {
-            printf("libarchive error : '%s'\n", archive_error_string(a));
             return false;
         }
     }
