+++ src/libui_sdl/main.cpp
@@ -34,7 +34,6 @@
 #include "DlgWifiSettings.h"
 
 #include "../NDS.h"
-#include "../NDSCart.h"
 #include "../GPU.h"
 #include "../SPU.h"
 #include "../Wifi.h"
@@ -1341,7 +1340,7 @@ void OnDropFile(uiWindow* window, char* file, void* blarg)
     char* ext = &file[strlen(file)-3];
     int prevstatus = EmuRunning;
 
-    if (!strcasecmp(ext, "nds") || !strcasecmp(ext, "srl") || NDSCart::CheckArchiveExtensions(ext))
+    if (!strcasecmp(ext, "nds") || !strcasecmp(ext, "srl") || Platform::CheckArchiveExtensions(ext))
     {
         if (RunningSomething)
         {
@@ -2062,7 +2061,7 @@ int main(int argc, char** argv)
         char* file = argv[1];
         char* ext = &file[strlen(file)-3];
 
-        if (!strcasecmp(ext, "nds") || !strcasecmp(ext, "srl") || NDSCart::CheckArchiveExtensions(ext))
+        if (!strcasecmp(ext, "nds") || !strcasecmp(ext, "srl") || Platform::CheckArchiveExtensions(ext))
         {
             strncpy(ROMPath, file, 1023);
             ROMPath[1023] = '\0';
